azure_functions/requirements.txt
azure-functions
psycopg2-binary
azure-identity
azure-storage-blob
pandas
xlsxwriter 
requests